Refactoring: voting text rendering
- Stop prerendering the vote texts (and only rerender them when needed).
- Move the vote-related hooks into vote-hooks.js (rather than app-logic.js).
- Move usePanelState() into utils-hooks.js.
