+++ apps/voting/app/src/components/VoteText.js
@@ -1,0 +1,42 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import { shortenAddress, transformAddresses } from '../web3-utils'
+import AutoLink from '../components/AutoLink'
+
+// Render a text associated to a vote.
+// Usually vote.data.metadata and vote.data.description.
+const VoteText = React.memo(
+  ({ text = '' }) => {
+    console.log('TEXT', text)
+    // If there is no text, the component doesn’t render anything.
+    if (!text.trim()) {
+      return null
+    }
+
+    return (
+      <AutoLink>
+        {text.split('\n').map((line, i) => (
+          <React.Fragment key={i}>
+            {transformAddresses(line, (part, isAddress, index) =>
+              isAddress ? (
+                <span title={part} key={index}>
+                  {shortenAddress(part)}
+                </span>
+              ) : (
+                <span key={index}>{part}</span>
+              )
+            )}
+            <br />
+          </React.Fragment>
+        ))}
+      </AutoLink>
+    )
+  },
+  (prevProps, nextProps) => prevProps.text === nextProps.text
+)
+
+VoteText.propTypes = {
+  text: PropTypes.string,
+}
+
+export default VoteText
