+++ apps/voting/app/src/utils-hooks.js
@@ -1,4 +1,5 @@
-import { useState, useEffect } from 'react'
+import { useCallback, useEffect, useState } from 'react'
+import { noop } from './utils'
 
 export function useNow(updateEvery = 1000) {
   const [now, setNow] = useState(new Date())
@@ -29,4 +30,40 @@ export function usePromise(fn, memoParams, defaultValue) {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [...memoParams, fn])
   return result
+}
+
+// Handles the state of a panel.
+// Pass `onTransitionEnd` to the same SidePanel prop.
+export function usePanelState({ onDidOpen = noop, onDidClose = noop } = {}) {
+  const [visible, setVisible] = useState(false)
+
+  // `didOpen` is set to `true` when the opening transition of the panel has
+  // ended, `false` otherwise. This is useful to know when to start inner
+  // transitions in the panel content.
+  const [didOpen, setDidOpen] = useState(false)
+
+  const requestOpen = useCallback(() => {
+    setVisible(true)
+    setDidOpen(false)
+  }, [setVisible, setDidOpen])
+
+  const requestClose = useCallback(() => {
+    setVisible(false)
+  }, [setVisible])
+
+  // To be passed to the onTransitionEnd prop of SidePanel.
+  const onTransitionEnd = useCallback(
+    opened => {
+      if (opened) {
+        onDidOpen()
+        setDidOpen(true)
+      } else {
+        onDidClose()
+        setDidOpen(false)
+      }
+    },
+    [onDidClose, onDidOpen, setDidOpen]
+  )
+
+  return { requestOpen, requestClose, visible, didOpen, onTransitionEnd }
 }
